+++ src/Crabtopus/CardManager.cs
@@ -7,7 +7,7 @@ using System.Net.Http;
 using System.Text;
 using System.Text.Json;
 using System.Threading.Tasks;
-using Crabtopus.Model;
+using Crabtopus.Models;
 
 namespace Crabtopus
 {
