+++ src/Crabtopus/LogReader.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
-using Crabtopus.Model;
+using Crabtopus.Models;
 
 namespace Crabtopus
 {
