+++ src/Crabtopus/Models/Asset.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace Crabtopus.Model
+namespace Crabtopus.Models
 {
     internal class Asset
     {
