+++ src/Crabtopus/Models/Card.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace Crabtopus.Model
+namespace Crabtopus.Models
 {
     internal class Card : IEquatable<Card>
     {
