+++ src/Crabtopus/Models/Wildcards.cs
@@ -1,7 +1,7 @@
 ﻿
 using System.Text.Json.Serialization;
 
-namespace Crabtopus.Model
+namespace Crabtopus.Models
 {
     internal class Wildcards
     {
